Move test checking repo code to checker package
